LayoutTests/editing/inserting/insert-list-in-table-cell-08-expected.txt

Exec insertOrderedList twice in a table removes the previously inserted list item:

Before:
| <#selection-anchor>
| <table>
|   border="1"
|   id="element"
|   <tbody>
|     <tr>
|       <td>
|         "fsdf"
|       <td>
|         "fsdf"
|     <tr>
|       <td>
|         "gghfg"
|       <td>
|         "fsfg"
| <#selection-focus>

After:
| <#selection-anchor>
| <table>
|   border="1"
|   id="element"
|   <tbody>
|     <tr>
|       <td>
|         "fsdf"
|       <td>
|         "fsdf"
|     <tr>
|       <td>
|         "gghfg"
|       <td>
|         "fsfg"
| <#selection-focus>
